+++ Index.html
@@ -30,7 +30,7 @@
 <!-- <div class="modal"></div> -->
 </div>
 
-<script type="module" src="./main.js"></script>
+<script type="module" src="./main.ts"></script>
 </body>
 
 </html>
