+++ tsconfig.json
@@ -6,6 +6,7 @@
       "module": "es6",
       "target": "es5",
       "allowJs": true,
+      "sourceMap": true,
       "moduleResolution": "node"
     }
   }
